Code changes for fdo#74107:File Corruption - Issue with outerShdw
Problem Description:
- In document.xml, outerShdw dist contains the negative value after roundtrip.
- Due to negative value file get corrupted.

XML difference:
 Roundtrip: Before fix
 <a:outerShdw dist="-1363652116480" dir="2700000">

 Roundtrip: After fix
 <a:outerShdw dist="57811035" dir="2700000">

Conflicts:
	sw/qa/extras/ooxmlexport/ooxmlexport.cxx
Reviewed on:
	https://gerrit.libreoffice.org/7880

+++ sw/source/filter/ww8/docxsdrexport.cxx
@@ -571,7 +571,7 @@ void DocxSdrExport::writeDMLEffectLst(const SwFrmFmt& rFrmFmt)
     if (aShadowItem.GetLocation() != SVX_SHADOW_NONE)
     {
         // Distance is measured diagonally from corner
-        double nShadowDist = sqrt((aShadowItem.GetWidth()*aShadowItem.GetWidth())*2.0);
+        double nShadowDist = sqrt((double)aShadowItem.GetWidth()*aShadowItem.GetWidth()*2.0);
         OString aShadowDist(OString::number(TwipsToEMU(nShadowDist)));
         OString aShadowColor = msfilter::util::ConvertColor(aShadowItem.GetColor());
         OString aShadowAlpha = lcl_ConvertTransparency(aShadowItem.GetColor());
